Organized Button Routes for stream/posts Folder
