+++ src/components/stream/Stream.jsx
@@ -5,7 +5,7 @@ import { withRouter } from "react-router-dom";
 import { Container, Columns, Section } from "react-bulma-components";
 
 import PostList from "./posts/PostList";
-import PopupModal from "./posts/PopupModal";
+import PopupModal from "./posts/modals/PopupModal";
 
 import { buttonLayerContainerStyle, streamLayerContainerStyle, newPostButtonStyle, pageStyle } from "../../styling/StyleComponents";
 import { createPost, updatePost, sharePost, likePost, retrieveLoggedInAuthorPosts, deletePost, retrieveInbox, retrieveAllAuthors } from "./StreamActions";
