+++ src/components/stream/posts/ConfirmButton.jsx
@@ -2,9 +2,9 @@
     props.action - onClick event
 */
 
-import { ConfirmButton } from "./postSVG";
+import { ConfirmButton } from "./buttons/postSVG";
 import { Container } from "react-bulma-components";
-import {color} from "./styling";
+import {color} from "../styling";
 
 
 // local styling
