+++ src/components/stream/posts/EditPostForm.jsx
@@ -5,12 +5,12 @@ import Markdown from "react-markdown";
 import ReactTags from "react-tag-autocomplete";
 import "./react-tags.css";
 
-import CancelButton from "./CancelButton";
-import ConfirmButton from "./ConfirmButton";
-import DeleteButton from "./DeleteButton";
+import CancelButton from "./buttons/CancelButton";
+import ConfirmButton from "./buttons/ConfirmButton";
+import DeleteButton from "./buttons/DeleteButton";
 import CheckBox from "./CheckBox";
 
-import { TextIcon, ImageIcon, MarkdownIcon, ToolTipIcon,  } from "./postSVG";
+import { TextIcon, ImageIcon, MarkdownIcon, ToolTipIcon,  } from "./buttons/postSVG";
 import { ImageUploadIcon } from "../../../styling/svgIcons";
 import { color } from "./styling";
 import PostTab from "./PostTab";
