+++ src/components/stream/posts/Post.jsx
@@ -9,10 +9,10 @@ import { withRouter } from "react-router-dom";
 import { Card, Content, Container, Button } from "react-bulma-components";
 import Dividor from "./Dividor";
 import { Link } from "react-router-dom";
-import EditButton from "./EditButton";
-import ShareButton from "./ShareButton";
-import LikeButton from "./LikeButton";
-import LikedButton from "./LikedButton";
+import EditButton from "./buttons/EditButton";
+import ShareButton from "./buttons/ShareButton";
+import LikeButton from "./buttons/LikeButton";
+import LikedButton from "./buttons/LikedButton";
 import { DescriptionStyle, dividorStyle, postStyle, categoriesStyle, signatureStyle, postContainerStyle,postTitleStyle, postContentStyle, footerButtonLayoutStyle } from "../../../styling/StyleComponents";
 import { color } from "./styling";
 // import { connectAdvanced } from "react-redux";
