+++ src/components/stream/posts/SelectedPost.jsx
@@ -4,10 +4,10 @@ import { connect } from "react-redux";
 import { withRouter } from "react-router-dom";
 import { Modal, Button, Content, Section, Card,Container,Columns, List } from "react-bulma-components";
 import Dividor from "./Dividor";
-import EditButton from "./EditButton";
-import ShareButton from "./ShareButton";
-import LikeButton from "./LikeButton";
-import LikedButton from "./LikedButton";
+import EditButton from "./buttons/EditButton";
+import ShareButton from "./buttons/ShareButton";
+import LikeButton from "./buttons/LikeButton";
+import LikedButton from "./buttons/LikedButton";
 import { dividorStyle, pageStyle,buttonLayerContainerStyle,newPostButtonStyle, streamLayerContainerStyle, postContainerStyle, postStyle, signatureStyle, postTitleStyle, DescriptionStyle, contentStyle, postContentStyle, categoriesStyle, footerButtonLayoutStyle} from "../../../styling/StyleComponents";
 import { color } from "../../../styling/ColorFontConfig";
 import InboxModalPopUp from "../../inbox/InboxModalPopUp";
