+++ src/components/stream/posts/buttons/EditButton.jsx
@@ -2,28 +2,27 @@
     props.action - onClick event
 */
 
-import { DeleteIcon } from "./buttons/postSVG";
+import { PencilIcon } from "./postSVG";
 import { Container } from "react-bulma-components";
 import {color} from "../styling";
 
 
 // local styling
 var buttonStyle = {
-    scale: "40",
+    scale: "20",
     style: {
-        padding: "1.2em",
-        fill: color.buttonRed,
+        fill: color.buttonGreen,
     }
 }
 
 // component
-function DeleteButton(props) {
+function EditButton(props) {
 
     return (
         <Container style={buttonStyle.style}>
-            <a onClick={props.action}><DeleteIcon svgScale={buttonStyle.scale}/></a>
+            <a onClick={props.action}><PencilIcon svgScale={buttonStyle.scale}/></a>
         </Container>
     );
 }
 
-export default DeleteButton;
+export default EditButton;
