+++ src/components/stream/posts/buttons/NewPostFloatingButton.jsx
