+++ src/components/stream/posts/buttons/ShareButton.jsx
@@ -4,7 +4,7 @@
 
 import { SharingIcon } from "./postSVG";
 import { Container } from "react-bulma-components";
-import {color} from "./styling";
+import {color} from "../styling";
 
 
 // local styling
